Minimap - ignore clicks when multi-selecting
