+++ apps/client/src/app/probable-waffle/game/core/probable-waffle.scene.ts
@@ -14,6 +14,8 @@ import {
   ProbableWaffleSpectatorData
 } from "@fuzzy-waddle/api-interfaces";
 import { ProbableWaffleCommunicatorService } from "../../communicators/probable-waffle-communicator.service";
+import { BehaviorSubject } from "rxjs";
+import { ProbableWaffleSceneData } from "../scenes/GameProbableWaffleScene";
 
 export class ProbableWaffleScene extends BaseScene<
   ProbableWaffleGameData,
@@ -28,6 +30,21 @@ export class ProbableWaffleScene extends BaseScene<
   ProbableWaffleSpectator,
   ProbableWaffleCommunicatorService
 > {
+  override getSceneGameData() {
+    return this.sceneGameData;
+  }
+
+  protected sceneGameData: ProbableWaffleSceneData = {
+    baseGameData: this.baseGameData,
+    systems: [],
+    components: [],
+    services: [],
+    initializers: {
+      sceneInitialized: new BehaviorSubject<boolean>(false),
+      postSceneInitialized: new BehaviorSubject<boolean>(false)
+    }
+  } satisfies ProbableWaffleSceneData;
+
   get mapInfo(): ProbableWaffleMapData {
     return ProbableWaffleLevels[this.baseGameData.gameInstance.data.gameModeData!.map!];
   }
