+++ apps/client/src/app/probable-waffle/game/scenes/components/scene-component-helpers.ts
@@ -1,7 +1,7 @@
-import GameProbableWaffleScene from "../GameProbableWaffleScene";
+import { ProbableWaffleScene } from "../../core/probable-waffle.scene";
 
 export function getSceneComponent<T>(scene: Phaser.Scene, componentClass: new (...args: any[]) => T): T | undefined {
-  if (!(scene instanceof GameProbableWaffleScene)) throw new Error("Scene is not of type GameProbableWaffleSceneData");
+  if (!(scene instanceof ProbableWaffleScene)) throw new Error("Scene is not of type ProbableWaffleSceneData");
   const component = scene.getSceneGameData().components.find((c) => c instanceof componentClass);
   if (!component) {
     return undefined;
@@ -10,7 +10,7 @@ export function getSceneComponent<T>(scene: Phaser.Scene, componentClass: new (.
 }
 
 export function getSceneService<T>(scene: Phaser.Scene, serviceClass: new (...args: any[]) => T): T | undefined {
-  if (!(scene instanceof GameProbableWaffleScene)) throw new Error("Scene is not of type GameProbableWaffleSceneData");
+  if (!(scene instanceof ProbableWaffleScene)) throw new Error("Scene is not of type ProbableWaffleSceneData");
   const service = scene.getSceneGameData().services.find((s) => s instanceof serviceClass);
   if (!service) {
     return undefined;
@@ -19,7 +19,7 @@ export function getSceneService<T>(scene: Phaser.Scene, serviceClass: new (...ar
 }
 
 export function getSceneSystem<T>(scene: Phaser.Scene, systemClass: new (...args: any[]) => T): T | undefined {
-  if (!(scene instanceof GameProbableWaffleScene)) throw new Error("Scene is not of type GameProbableWaffleSceneData");
+  if (!(scene instanceof ProbableWaffleScene)) throw new Error("Scene is not of type ProbableWaffleSceneData");
   const system = scene.getSceneGameData().systems.find((s) => s instanceof systemClass);
   if (!system) {
     return undefined;
@@ -28,6 +28,6 @@ export function getSceneSystem<T>(scene: Phaser.Scene, systemClass: new (...args
 }
 
 export function getSceneInitializers(scene: Phaser.Scene) {
-  if (!(scene instanceof GameProbableWaffleScene)) throw new Error("Scene is not of type GameProbableWaffleSceneData");
+  if (!(scene instanceof ProbableWaffleScene)) throw new Error("Scene is not of type ProbableWaffleSceneData");
   return scene.getSceneGameData().initializers;
 }
